+++ lib/cldr/gettext/gettext.ex
@@ -1,4 +1,4 @@
-if Code.ensure_loaded?(Gettext) do
+if Mix.env in [:dev, :test] do
   defmodule Cldr.Gettext do
     @moduledoc """
     Implements a Gettext-compatible module but using Cldr locales.  Its an example