+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# Changelog for Cldr v1.3.1
+# Changelog for Cldr v1.3.2
 
-This is the changelog for Cldr v1.3.0 released on January 19th, 2018.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr/tags)
+This is the changelog for Cldr v1.3.0 released on January 20th, 2018.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/cldr/tags)
+
+### Bug Fixes
+
+* Correctly retrieve the default locale from a configured Gettext backend module.  Fixes #32.
+
+# Changelog for Cldr v1.3.1
 
 ### Bug Fixes